// generated/models/ContentDiscoveryBookWithSubject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Learning API
 * API with password OAuth2 authentication
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface ContentDiscoveryBookWithSubject
 */
export interface ContentDiscoveryBookWithSubject {
    /**
     *
     * @type {string}
     * @memberof ContentDiscoveryBookWithSubject
     */
    authors?: string;
    /**
     *
     * @type {string}
     * @memberof ContentDiscoveryBookWithSubject
     */
    description?: string;
    /**
     *
     * @type {string}
     * @memberof ContentDiscoveryBookWithSubject
     */
    pdfLink?: string;
    /**
     *
     * @type {boolean}
     * @memberof ContentDiscoveryBookWithSubject
     */
    selected?: boolean;
    /**
     *
     * @type {string}
     * @memberof ContentDiscoveryBookWithSubject
     */
    subjectId?: string;
    /**
     *
     * @type {string}
     * @memberof ContentDiscoveryBookWithSubject
     */
    subjectName?: string;
    /**
     *
     * @type {string}
     * @memberof ContentDiscoveryBookWithSubject
     */
    subjectUrl?: string;
    /**
     *
     * @type {string}
     * @memberof ContentDiscoveryBookWithSubject
     */
    title?: string;
    /**
     *
     * @type {string}
     * @memberof ContentDiscoveryBookWithSubject
     */
    url?: string;
}

/**
 * Check if a given object implements the ContentDiscoveryBookWithSubject interface.
 */
export function instanceOfContentDiscoveryBookWithSubject(value: object): value is ContentDiscoveryBookWithSubject {
    return true;
}

export function ContentDiscoveryBookWithSubjectFromJSON(json: any): ContentDiscoveryBookWithSubject {
    return ContentDiscoveryBookWithSubjectFromJSONTyped(json, false);
}

export function ContentDiscoveryBookWithSubjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContentDiscoveryBookWithSubject {
    if (json == null) {
        return json;
    }
    return {
        authors: json["authors"] == null ? undefined : json["authors"],
        description: json["description"] == null ? undefined : json["description"],
        pdfLink: json["pdf_link"] == null ? undefined : json["pdf_link"],
        selected: json["selected"] == null ? undefined : json["selected"],
        subjectId: json["subject_id"] == null ? undefined : json["subject_id"],
        subjectName: json["subject_name"] == null ? undefined : json["subject_name"],
        subjectUrl: json["subject_url"] == null ? undefined : json["subject_url"],
        title: json["title"] == null ? undefined : json["title"],
        url: json["url"] == null ? undefined : json["url"],
    };
}

export function ContentDiscoveryBookWithSubjectToJSON(json: any): ContentDiscoveryBookWithSubject {
    return ContentDiscoveryBookWithSubjectToJSONTyped(json, false);
}

export function ContentDiscoveryBookWithSubjectToJSONTyped(value?: ContentDiscoveryBookWithSubject | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        authors: value["authors"],
        description: value["description"],
        pdf_link: value["pdfLink"],
        selected: value["selected"],
        subject_id: value["subjectId"],
        subject_name: value["subjectName"],
        subject_url: value["subjectUrl"],
        title: value["title"],
        url: value["url"],
    };
}